dotgit: Move package outside internal.
