+++ storage/filesystem/module.go
@@ -2,7 +2,7 @@ package filesystem
 
 import (
 	"gopkg.in/src-d/go-git.v4/storage"
-	"gopkg.in/src-d/go-git.v4/storage/filesystem/internal/dotgit"
+	"gopkg.in/src-d/go-git.v4/storage/filesystem/dotgit"
 )
 
 type ModuleStorage struct {
