+++ storage/filesystem/reference.go
@@ -3,7 +3,7 @@ package filesystem
 import (
 	"gopkg.in/src-d/go-git.v4/plumbing"
 	"gopkg.in/src-d/go-git.v4/plumbing/storer"
-	"gopkg.in/src-d/go-git.v4/storage/filesystem/internal/dotgit"
+	"gopkg.in/src-d/go-git.v4/storage/filesystem/dotgit"
 )
 
 type ReferenceStorage struct {
