+++ storage/filesystem/storage.go
@@ -2,7 +2,7 @@
 package filesystem
 
 import (
-	"gopkg.in/src-d/go-git.v4/storage/filesystem/internal/dotgit"
+	"gopkg.in/src-d/go-git.v4/storage/filesystem/dotgit"
 
 	"gopkg.in/src-d/go-billy.v4"
 )
